+++ app/javascript/Components/course_summaries_table.jsx
@@ -84,11 +84,6 @@ export class CourseSummaryTable extends React.Component {
   };
 
   render() {
-    const effectiveLoadingComponent =
-      process.env.NODE_ENV === "test"
-        ? this.props.LoadingComponent
-        : ReactTableDefaults.LoadingComponent;
-
     return [
       !this.props.student && (
         <div key="show-hidden" style={{height: "2em"}}>
@@ -115,7 +110,6 @@ export class CourseSummaryTable extends React.Component {
           },
         ]}
         loading={this.props.loading}
-        LoadingComponent={effectiveLoadingComponent}
         filtered={this.state.filtered}
         onFilteredChange={filtered => this.setState({filtered})}
         className={"auto-overflow"}