+++ app/javascript/Components/notes_table.jsx
@@ -112,18 +112,12 @@ class NotesTable extends React.Component {
   ];
 
   render() {
-    const effectiveLoadingComponent =
-      process.env.NODE_ENV === "test"
-        ? this.props.LoadingComponent
-        : ReactTableDefaults.LoadingComponent;
-
     return (
       <ReactTable
         data={this.data()}
         columns={this.columns}
         sortable={false}
         loading={this.state.loading}
-        LoadingComponent={effectiveLoadingComponent}
         getNoDataProps={() => ({
           loading: this.state.loading,
         })}